Created a branch to refactor the code to add another table and create a relation between them.
The intention was to create a beans table with a foreign key in the reviews table (because there can be multiple reviews of one bean prepared several ways. However, I tried to do this using he exisiting user input to create two tables simultaniously. SQL did not like this, I then tried triggers in SQL, this was also too complicated.

I think the correct way to do this might be to create one table (beans) after the first input, and then create the review after the next two inputs.